azure-pipeline/requirements.txt
azure-functions>=1.15.0
azure-storage-blob>=12.17.0
azure-identity>=1.13.0
azure-storage-file-datalake>=12.12.0
azure-cosmos>=4.5.1
azure-search-documents>=11.4.0
azure-eventgrid>=4.12.0
azure-keyvault-secrets>=4.7.0
azure-mgmt-databricks>=1.1.0
azure-openai>=1.0.0
beautifulsoup4>=4.12.2
feedparser>=6.0.10
lxml>=4.9.3
pandas>=2.1.1
pytz>=2023.3
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.5
numpy>=1.24.3
tqdm>=4.66.1
